requirements.txt
# WebSocket & HTTP
websockets==12.0
aiohttp==3.9.5

# Dashboard (FastAPI)
fastapi==0.115.5
uvicorn==0.32.1
jinja2==3.1.4

# Data Processing
pandas==2.2.2
numpy==1.26.4
scipy==1.11.4

# Technical Analysis
ta==0.11.0

# Database
asyncpg==0.29.0

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.1

# Monitoring
prometheus-client==0.19.0

# AI Integration
anthropic==0.18.1
psutil==5.9.8